+++ src/demo/components/layout/header/StickyHeader.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { Header, StickyHeader } from "../../../../index.js";
 
 const LongContent: React.FC = () => {
-    // eslint-disable-next-line @typescript-eslint/no-magic-numbers
+    // eslint-disable-next-line @typescript-eslint/no-magic-numbers,react/no-array-index-key
     const longContent = new Array(1000).fill(null).map((_, key) => <div key={key}>content</div>);
     return <>{longContent}</>;
 };
@@ -11,7 +11,7 @@ const LongContent: React.FC = () => {
 const StickyHeaderDemo = () => {
     return (
         <StickyHeader>
-            <Header >Some place</Header>
+            <Header>Some place</Header>
             <StickyHeader.Content>
                 <LongContent />
             </StickyHeader.Content>