+++ src/pages/_app.tsx
@@ -1,11 +1,17 @@
 import React from "react";
+import Head from "next/head";
 import type { AppProps } from "next/app";
 
 import "../global.scss";
 import "../demo-global.scss";
 
 const MyApp = ({ Component, pageProps }: AppProps) => (
-    <Component {...pageProps} />
+    <>
+        <Head>
+            <title>MIUI in React</title>
+        </Head>
+        <Component {...pageProps} />
+    </>
 );
 
 export default MyApp;